Apply first round of static analysis fixes.
This commit applies some trivial static analysis fixes for the
Groningen codebase, many of which relate to Java Language Level
Seven:

1.) Generics type inferences.
2.) Unused imports.
3.) Finalization of fields.
4.) Collapsing of identical catch statements into multi-catch.
5.) JUnit test name conventions.
6.) Removal of explicit autoboxing.

+++ src/main/java/org/arbeitspferde/groningen/extractor/Extractor.java
@@ -25,7 +25,6 @@ import org.arbeitspferde.groningen.utility.MetricExporter;
 
 import javax.annotation.Nullable;
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
@@ -152,7 +151,7 @@ public class Extractor implements Runnable {
               double pauseTimeSecs;
 
               try {
-                pauseTimeSecs = Double.valueOf(token.trim()).doubleValue();
+                pauseTimeSecs = Double.valueOf(token.trim());
               } catch (final Exception e) {
                 log.log(Level.WARNING,
                     String.format("Unable to parse pause time '%s'. Defaulting to 0.0", token), e);
